Move mapper logic.
